added helper mixin for total count logic
